refactor: cupti_backend refactoring
