+++ CMakeLists.txt
@@ -257,7 +257,12 @@ if(GPUFL_ENABLE_NVIDIA)
                 include/gpufl/backends/nvidia/sampler/cupti_sass.cpp
                 include/gpufl/backends/nvidia/sampler/cupti_sass.hpp
                 include/gpufl/backends/nvidia/cuda_collector.cpp
+                include/gpufl/backends/nvidia/cupti_activity_callbacks.cpp
+                include/gpufl/backends/nvidia/cupti_activity_state.cpp
                 include/gpufl/backends/nvidia/cupti_utils.cpp
+                include/gpufl/backends/nvidia/cupti_capture_capabilities.cpp
+                include/gpufl/backends/nvidia/cupti_engine_selection.cpp
+                include/gpufl/backends/nvidia/cupti_runtime_support.cpp
                 include/gpufl/backends/nvidia/cuda_cleanup_handler.cpp
                 include/gpufl/backends/nvidia/resource_handler.cpp
                 include/gpufl/backends/nvidia/kernel_launch_handler.cpp
